+++ src/components/common/Select.ts
@@ -1,211 +1,13 @@
-import { Select, useUniqueId } from '@terra-ui-packages/components';
+import { Select } from '@terra-ui-packages/components';
 import _ from 'lodash/fp';
 import { Children, useCallback, useEffect, useRef } from 'react';
 import { div, h } from 'react-hyperscript-helpers';
-import RSelect, {
-  components as RSelectComponents,
-  GroupBase as RSelectGroupBase,
-  MultiValue as RSelectMultiValue,
-  Props as RSelectProps,
-  PropsValue as RSelectPropsValue,
-  SingleValue as RSelectSingleValue,
-} from 'react-select';
-import RAsyncCreatableSelect from 'react-select/async-creatable';
 import { List } from 'react-virtualized';
 import { AutoSizer } from 'src/components/common/VirtualizedSelectAutoSizer';
-import { icon } from 'src/components/icons';
-import colors from 'src/libs/colors';
 import { useOnMount } from 'src/libs/react-utils';
 
-export { Select } from '@terra-ui-packages/components';
-
-const commonSelectProps: Partial<RSelectProps> = {
-  theme: (base) =>
-    _.merge(base, {
-      colors: {
-        primary: colors.accent(),
-        neutral20: colors.dark(0.55),
-        neutral30: colors.dark(0.55),
-      },
-      spacing: { controlHeight: 36 },
-    }),
-  styles: {
-    control: (base, { isDisabled }) =>
-      _.merge(base, {
-        backgroundColor: isDisabled ? colors.dark(0.25) : 'white',
-        boxShadow: 'none',
-      }),
-    singleValue: (base) => ({ ...base, color: colors.dark() }),
-    option: (base, { isSelected, isFocused, isDisabled }) =>
-      _.merge(base, {
-        fontWeight: isSelected ? 600 : undefined,
-        backgroundColor: isFocused ? colors.dark(0.15) : 'white',
-        color: isDisabled ? undefined : colors.dark(),
-        ':active': { backgroundColor: colors.accent(isSelected ? 0.55 : 0.4) },
-      }),
-    clearIndicator: (base) => ({ ...base, paddingRight: 0 }),
-    indicatorSeparator: () => ({ display: 'none' }),
-    dropdownIndicator: (base, { selectProps: { isClearable } }) =>
-      _.merge(base, { paddingLeft: isClearable ? 0 : undefined }),
-    multiValueLabel: (base) => ({ ...base, maxWidth: '100%' }),
-    multiValueRemove: (base) => _.merge(base, { ':hover': { backgroundColor: 'unset' } }),
-    placeholder: (base) => ({ ...base, color: colors.dark(0.8) }),
-  },
-  components: {
-    Option: ({ children, selectProps, ...props }) =>
-      h(
-        RSelectComponents.Option,
-        _.merge(props, {
-          selectProps,
-          innerProps: {
-            role: 'option',
-            'aria-selected': props.isSelected,
-          },
-        }),
-        [
-          div({ style: { display: 'flex', alignItems: 'center', minHeight: 25 } }, [
-            div({ style: { flex: 1, minWidth: 0, overflowWrap: 'break-word' } }, [children]),
-            props.isSelected &&
-              icon('check', { size: 14, style: { flex: 'none', marginLeft: '0.5rem', color: colors.dark(0.5) } }),
-          ]),
-        ]
-      ),
-    Menu: ({ children, selectProps, ...props }) =>
-      h(
-        RSelectComponents.Menu,
-        _.merge(props, {
-          selectProps,
-          innerProps: {
-            role: 'listbox',
-            'aria-label': 'Options',
-            'aria-multiselectable': selectProps.isMulti,
-          },
-        }),
-        [children]
-      ),
-  },
-};
-
-const formatGroupLabel = (group: { label?: string }) =>
-  div(
-    {
-      style: {
-        color: colors.dark(),
-        fontSize: 14,
-        height: 30,
-        fontWeight: 600,
-        borderBottom: `1px solid ${colors.dark(0.25)}`,
-      },
-    },
-    [group.label]
-  );
-
-type BaseSelectProps<
-  Value,
-  Option extends { value: Value; label?: string | undefined },
-  IsMulti extends boolean,
-  Group extends RSelectGroupBase<Option>
-> = Omit<RSelectProps<Option, IsMulti, Group>, 'getOptionLabel' | 'value'> & {
-  findValue: (value: Value) => Option | null;
-  // react-select requires getOptionLabel. We provide a default, so it's optional.
-  getOptionLabel?: (option: Option) => string;
-  // react-select's type for value does not take IsMulti into account.
-  value: IsMulti extends true ? RSelectMultiValue<Value> : RSelectSingleValue<Value>;
-};
-
-/**
- * Wraps RSelect.
- *
- * Provides defaults for formatGroupLabel, getOptionLabel, and inputId.
- *
- * Also (via findValue) allows passing value as the selected value instead of the selected option.
- * For example:
- * const options = [{ value: 'foo' }, { value: 'bar' }]
- * h(BaseSelect, { options, value: 'foo' })
- * instead of:
- * const options = [{ value: 'foo' }, { value: 'bar' }]
- * h(BaseSelect, { options, value: options[0] })
- */
-const BaseSelect = <
-  Value,
-  Option extends { value: Value; label?: string | undefined },
-  IsMulti extends boolean,
-  Group extends RSelectGroupBase<Option>
->({
-  value,
-  id,
-  findValue,
-  ...props
-}: BaseSelectProps<Value, Option, IsMulti, Group>) => {
-  const newValue: RSelectPropsValue<Option> = props.isMulti
-    ? _.compact(_.map(findValue, value as Value[]))
-    : findValue(value as Value);
-  const myId = useUniqueId();
-  const inputId = id || myId;
-
-  const ParameterizedRSelect = RSelect as typeof RSelect<Option, IsMulti, Group>;
-  return h(
-    ParameterizedRSelect,
-    _.merge(
-      {
-        inputId,
-        ...commonSelectProps,
-        getOptionLabel: ({ value, label }) => label || value.toString(),
-        value: newValue || null, // need null instead of undefined to clear the select
-        formatGroupLabel,
-      },
-      props
-    )
-  );
-};
-
-export type SelectProps<
-  Value,
-  IsMulti extends boolean,
-  Option extends { value: Value; label?: string | undefined }
-> = Omit<BaseSelectProps<Value, Option, IsMulti, never>, 'findValue' | 'options'> & {
-  options: (Value | Option)[];
-};
-
-export type GroupedSelectProps<
-  Value,
-  IsMulti extends boolean,
-  Option extends { value: Value; label?: string | undefined },
-  Group extends RSelectGroupBase<Option>
-> = Omit<BaseSelectProps<Value, Option, IsMulti, Group>, 'findValue' | 'options'> & {
-  options: Group[];
-};
-
-/**
- * @param props - see {@link https://react-select.com/props#select-props}
- * @param props.value - a member of an inner options object
- * @param props.options - an object with toplevel pairs of label:options where label is a group label and options is an array of objects containing value:label pairs
- * @param [props.id] - The HTML ID to give the form element
- */
-export const GroupedSelect = <
-  Value,
-  IsMulti extends boolean = false,
-  Option extends { value: Value; label?: string | undefined } = { value: Value; label?: string | undefined },
-  Group extends RSelectGroupBase<Option> = RSelectGroupBase<Option>
->({
-  value,
-  options = [],
-  ...props
-}: GroupedSelectProps<Value, IsMulti, Option, Group>) => {
-  // cast because types don't carry through Lodash
-  const flattenedOptions = _.flatMap('options', options) as Option[];
-  const findValue = (target: Value) => (_.find({ value: target }, flattenedOptions) || null) as Option | null;
-
-  const ParameterizedBaseSelect = BaseSelect as typeof BaseSelect<Value, Option, IsMulti, Group>;
-  return h(ParameterizedBaseSelect, { value, options, findValue, ...props });
-};
-
-export const AsyncCreatableSelect = (props) => {
-  return h(RAsyncCreatableSelect, {
-    ...commonSelectProps,
-    ...props,
-  });
-};
+export { AsyncCreatableSelect, GroupedSelect, Select } from '@terra-ui-packages/components';
+export type { SelectProps } from '@terra-ui-packages/components';
 
 const VirtualizedMenuList = (props) => {
   const { children, focusedOption, getStyles, getValue, maxHeight } = props;
